Make external and fip network configurable
The floatingip_network was hard-coded to "public" in all the example
scenario heat templates. This change adds a new variable to make that
user configurable.

The bootstrap_vars.yaml files have the following new vars as example.

  os_floating_network: public
  os_router_external_network: public

Change those, or override on the commandline with -e <var_name> with
values appropriate for the cloud being used.