Add value_specs override for all networks
Operating on a cloud with bad configuration, incurrect MTU is causing a
massive amount of network traffic fragmentation. Result is 7-8x degraded
performance in network transfers, and a magnitude of unnececary CPU
load/interupts.

Adding the value_specs, so that the MTU can be forced for the networks
to use a better value.

ref: https://access.redhat.com/solutions/7059376