hotloop_yaml_patch - is already set
Check if value is already set, and skip replacing in this case.

Also ignore the where conditions in case the doc already has the
path and value defined in the patch. This allows for where conditions
to match on the patch path i.e to replace only if value is `null`.
NOTE: If the last key of the path is not defined, the lookup returns
`None`|`null`.

```yaml
- name: Patch value in YAML if key not present or has null value
  hotloop_yaml_patch:
    file: /tmp/hot_patch_test.yaml
    path: spec.nodeTemplate.ansible.ansibleVars.edpm_podman_registries
    value: REPLACED
    where:
      - path: spec.nodeTemplate.ansible.ansibleVars.edpm_podman_registries
        value: null
```